deal with message attribute in all controllers
always pass in to every get and post method
should I set it in every new model then? -- worried this will never
allow it to be empty.  Not setting in new models for now

also concerned how it works if the form adds a message attribute and
then in the controller I "add" a message attribute. Will it override the
first like I hope?