add Config class to list of RootConfigClasses
because "Configuring a DataSource in Spring MVC should be done one of
the classes returned from getRootConfigClasses"

The Root Configuration class is supposed to be the place where I
configure beans for my service and DAO layers.. which I think I've
already done in the Config class. I'm hoping adding it here will mean I
don't need to move anything to that RootConfiguration class.