close ApplicationContext after invoking DAO methods
this was the reason for the EntityManagerFactory closed exception
EMF actually wraps SessionFactory, so the issue was closing
SessionFactories
When you close an ApplicationContext, its SessionFactory bean gets
destroyed.  All beans get destroyed when the ApplicationContext is
closed, which makes sense.
moving these close methods prevents the need to make the SF bean
prototype scoped (which would create a new SF every time)